refactor: use TestRestOperateClient for container based tests with operate queries
